requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10
django-appconf==1.0.2
django-compressor==2.1
django-filepicker==0.2.2
django-markup==1.1
django-sass-processor==0.4.6
django-simple-captcha==0.5.2
django-storages==1.5.0
django-template-repl==0.3.0
gunicorn==19.6.0
httplib2==0.9.2
libsass==0.11.1
Markdown==2.6.6
oauth2client==3.0.0
Pillow==3.3.1
ply==3.8
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.5.3
rcssmin==1.0.6
requests==2.11.1
rjsmin==1.0.12
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
slimit==0.8.1
static3==0.7.0
uritemplate==0.6
